+++ R/class-export.R
@@ -12,6 +12,7 @@
 #' @slot unit `r get_slot_doc("export", "unit")`
 #' @slot reserve `r get_slot_doc("export", "reserve")`
 #' @slot export `r get_slot_doc("export", "export")`
+#' @slot region `r get_slot_doc("export", "region")`
 #' @slot misc `r get_slot_doc("export", "misc")`
 #'
 #' @include class-import.R
@@ -26,10 +27,10 @@ setClass("export",
     desc = "character",
     commodity = "character",
     unit = "character",
-    # !!! add @region
     reserve = "data.frame",
     export = "data.frame",
     cluster = "data.frame",
+    region = "character",
     # timeframe = "character", # depreciated (equal to commodity@timeframe)
     misc = "list"
   ),
@@ -89,6 +90,7 @@ setClass("export",
       order = integer(),
       stringsAsFactors = FALSE
     ),
+    region = character(),
     # ! Misc
     misc = list()
   ),
@@ -128,6 +130,7 @@ setMethod("initialize", "export", function(.Object, ...) {
 #' @param unit `r get_slot_doc("export", "unit")`
 #' @param reserve `r get_slot_doc("export", "reserve")`
 #' @param export `r get_slot_doc("export", "export")`
+#' @param region `r get_slot_doc("export", "region")`
 #' @param misc `r get_slot_doc("export", "misc")`
 #'
 #' @return export object with given specifications.
@@ -158,6 +161,7 @@ newExport <- function(
     reserve = data.frame(),
     export = data.frame(),
     cluster = data.frame(),
+    region = character(),
     misc = list(),
     ...
     ) {
@@ -176,6 +180,7 @@ newExport <- function(
     reserve = reserve,
     export = export,
     cluster = cluster,
+    region = region,
     misc = misc,
     ...
     )