+++ R/map_closure.R
@@ -242,6 +242,7 @@ map_mCommReg <- function(scen, fmp) {
   # checking here: it needs the collected process/commodity relations.
   .check_process_levels(scen)
 
+
   scen
 }